% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_data}
\alias{available_data}
\title{Data availability through SUNGEO API}
\format{
List of 42 data.table objects
Geoset:GADM :Classes ‘data.table’ and 'data.frame': 249 obs. of  4 variables
Geoset:GAUL :Classes ‘data.table’ and 'data.frame': 242 obs. of  4 variables
Geoset:geoBoundaries :Classes ‘data.table’ and 'data.frame': 197 obs. of  4 variables
Geoset:GRED :Classes ‘data.table’ and 'data.frame': 74 obs. of  4 variables
Geoset:HEXGRID :Classes ‘data.table’ and 'data.frame': 199 obs. of  4 variables
Geoset:MPIDR :Classes ‘data.table’ and 'data.frame': 52 obs. of  4 variables
Geoset:NHGIS :Classes ‘data.table’ and 'data.frame': 1 obs. of  4 variables
Geoset:PRIOGRID :Classes ‘data.table’ and 'data.frame': 199 obs. of  4 variables
Geoset:SHGIS :Classes ‘data.table’ and 'data.frame': 68 obs. of  4 variables
\describe{
  \item{country_iso3 }{Codes for available countries (ISO 3166-1 alpha-3). Character string.}
  \item{country_name }{Names of available countries. Character string.}
  \item{geoset_years }{Years of available historical boundary files. Character string.}
  \item{space_units }{Available spatial units of analysis. Character string.}
}
Elections:LowerHouse:CLEA :Classes ‘data.table’ and 'data.frame': 168 obs. of  6 variables
Demographics:Ethnicity:EPR :Classes ‘data.table’ and 'data.frame': 180 obs. of  6 variables
Demographics:Ethnicity:GREG :Classes ‘data.table’ and 'data.frame': 234 obs. of  6 variables
Demographics:Population:GHS :Classes ‘data.table’ and 'data.frame': 257 obs. of  6 variables
Events:PoliticalViolence:ABADarfur :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:ACLED :Classes ‘data.table’ and 'data.frame': 100 obs. of  6 variables
Events:PoliticalViolence:BeissingerProtest :Classes ‘data.table’ and 'data.frame': 15 obs. of  6 variables
Events:PoliticalViolence:BeissingerRiot :Classes ‘data.table’ and 'data.frame': 15 obs. of  6 variables
Events:PoliticalViolence:BeissingerUkraine :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:COCACW :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:ESOCAfghanistanWITS :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:ESOCIraqSIGACT :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:ESOCIraqWITS :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:ESOCMexicoDrugRelatedMurders :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:ESOCMexicoHomicide :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:ESOCPakistanBFRS :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:ESOCPakistanWITS :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:GED :Classes ‘data.table’ and 'data.frame': 121 obs. of  6 variables
Events:PoliticalViolence:Lankina :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:NIRI :Classes ‘data.table’ and 'data.frame': 12 obs. of  6 variables
Events:PoliticalViolence:NVMS :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:PITF :Classes ‘data.table’ and 'data.frame': 133 obs. of  6 variables
Events:PoliticalViolence:SCAD :Classes ‘data.table’ and 'data.frame': 60 obs. of  6 variables
Events:PoliticalViolence:yzCaucasus2000 :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:yzChechnya :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:yzLibya :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Events:PoliticalViolence:yzUkraine2014 :Classes ‘data.table’ and 'data.frame': 1 obs. of  6 variables
Infrastructure:Roads:gRoads :Classes ‘data.table’ and 'data.frame': 240 obs. of  6 variables
Infrastructure:NightLights:DMSP :Classes ‘data.table’ and 'data.frame': 257 obs. of  6 variables
PublicHealth:Covid19:JHUCSSEC19 :Classes ‘data.table’ and 'data.frame': 207 obs. of  6 variables
Terrain:Elevation:ETOPO1 :Classes ‘data.table’ and 'data.frame': 256 obs. of  6 variables
Terrain:LandCover:GLCC :Classes ‘data.table’ and 'data.frame': 257 obs. of  6 variables
Weather:AirTemperatureAndPrecipitation:NOAA :Classes ‘data.table’ and 'data.frame': 209 obs. of  6 variables
\describe{
  \item{country_iso3 }{Codes for available countries (ISO 3166-1 alpha-3). Character string.}
  \item{country_name }{Names of available countries. Character string.}
  \item{year_range }{Range of available years for data topic. Character string.}
  \item{time_units }{Available time units. Character string.}
  \item{space_units }{Available spatial units. Character string.}
  \item{geosets }{Names of available geographic boundary data sources. Character string.}
}
}
\source{
Sub-National Geospatial Data Archive System: Geoprocessing Toolkit (updated March 17, 2023).
}
\usage{
available_data
}
\description{
Census of geospatial and processed data files available to download using \code{SUNGEO::get_data()}.
}
\keyword{datasets}
