% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_dict}
\alias{cc_dict}
\title{Country code dictionary}
\format{
data.table object, with	8626 obs. of  3 variables:
\describe{
  \item{country_name }{Country names. Character string.}
  \item{country_name_alt }{Alternative spellings of country names, ASCII characters only. Character string.}
  \item{country_iso3 }{Country codes (ISO 3166-1 alpha-3). Character string.}
}
}
\source{
Vincent Arel-Bundock. Package countrycode: Convert Country Names and Country Code, version 1.40. CRAN (October 12, 2022).
}
\usage{
cc_dict
}
\description{
Reference table of country names and ISO-3166 codes, adapted from \code{countrycode} package.
}
\keyword{datasets}
