% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clea_deu2009}
\alias{clea_deu2009}
\title{Constituency level results for lower chamber legislative elections, Germany 2009.}
\format{
Simple feature collection with 16 features and 10 fields.
geometry type:  MULTIPOLYGON.
dimension:      XY.
bbox:           xmin: 5.867281 ymin: 47.27096 xmax: 15.04388 ymax: 55.05902.
epsg (SRID):    4326.
proj4string:    +proj=longlat +datum=WGS84 +no_defs.
\describe{
  \item{cst }{Constituency number. Numeric.}
  \item{cst_n }{Constituency name. Character.}
  \item{ctr }{Country number. Numeric.}
  \item{ctr_n }{Country name. Character.}
  \item{yrmo }{Year and month of election (YYYYMM). Character.}
  \item{to1 }{Turnout in first round. Numeric.}
  \item{vv1 }{Number of valid votes in first round. Numeric.}
  \item{pvs1_margin }{Popular vote share margin in first round. Numeric.}
  \item{incumb_pty_n }{Incumbent party name.}
  \item{win1_pty_n }{Party name of popular vote share winner in first round. Character.}
}
}
\source{
Constituency-Level Elections Archive (CLEA) \url{https://electiondataarchive.org/}
}
\usage{
clea_deu2009
}
\description{
A simple feature collection containing the spatial geometries of electoral constituency
borders, and data on turnout levels, votes shares and other attributes of lower chamber
legislative elections.
}
\keyword{datasets}
