% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_geonames.R
\name{clean_geonames}
\alias{clean_geonames}
\title{Clean GeoNames}
\usage{
clean_geonames(gn, ppl = FALSE)
}
\arguments{
\item{gn}{Raw GeoNames data file. \code{data.frame} or \code{data.table} object.}

\item{ppl}{Keep only populated place names? Logical.}
}
\value{
Returns a \code{data.table} with properly-formatted GeoNames information.
}
\description{
Function to clean GeoNames gazetteer files.
}
\details{
Function used internally by \code{geocode_gn}
}
