% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_select.R
\name{crs_select}
\alias{crs_select}
\title{Automatic planar coordinate reference system (CRS) selection}
\usage{
crs_select(polyz, sf_layer = polyz)
}
\arguments{
\item{polyz}{Polygon layer definiting the study region. \code{sf} object.}

\item{sf_layer}{Layer to be reprojected, if different from \code{polyz}. \code{sf} object.}
}
\value{
List object, containing
\itemize{
 \item{"sf". }{The re-projected layer. \code{sf} object.}
 \item{"epsg_best". }{EPSG code of the projection. Character string.}
 }
}
\description{
Function to automatically convert simple features with geodetic coordinates (longitude, latitude / WGS 1984, EPSG:4326) to planar coordinates.
}
\details{
Optimal map projection for the object \code{sf_layer} is defined as one that maximizes areal overlap between the study area and the spatial extent of planar coordinate reference systems and transformations in the EPSG Geodetic Parameter Dataset (polygons version 9.8, 2019-09-19).
}
\examples{
# Find a planar projection for an unprojected (WSG 1984) hexagonal grid of Germany
\dontrun{
data(hex_05_deu)
hex_tr <- crs_select(hex_05_deu)
}
}
