% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dcwroad_deu1992}
\alias{dcwroad_deu1992}
\title{Roads polylines for Germany, 1992}
\format{Simple feature collection with 4713 features and 5 fields.
geometry type:  MULTILINESTRING.
dimension:      XY.
bbox:           xmin: 5.750933 ymin: 47.30845 xmax: 14.99799 ymax: 55.01572.
epsg (SRID):    4326.
proj4string:    +proj=longlat +datum=WGS84 +no_defs.
\describe{
  \item{MED_DESCRI }{Is the road a divided multi-lane highway with a median? Character string.}
  \item{RTT_DESCRI }{Primary or secondary route? Character string.}
  \item{F_CODE_DES }{Feature code description (road or trail). Character string.}
  \item{ISO }{ISO 3166-1 alpha-3 country code. Character string.}
  \item{ISOCOUNTRY }{Country name. Character string.}
}}
\source{
Defense Mapping Agency (DMA), 1992. Digital Chart of the World. Defense Mapping Agency, Fairfax, Virginia. (Four CD-ROMs.)  \url{https://earth-info.nga.mil/publications/specs/printed/89009/89009_DCW.pdf}. Available through DIVA-GIS: \url{http://www.diva-gis.org/gData} (accessed April 15, 2020).
}
\usage{
dcwroad_deu1992
}
\description{
Roads thematic layer from Digital Chart of the World.
}
\keyword{datasets}
