% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_geom.R
\name{fix_geom}
\alias{fix_geom}
\title{Polygon geometry fix}
\usage{
fix_geom(x, self_int = TRUE)
}
\arguments{
\item{x}{Polygon layer to be checked and fixed. \code{sf} object.}

\item{self_int}{Look only for self-intersections? Logical.}
}
\value{
Returns a \code{sf} polygon object, with self-intersections and other geometry problems fixed.
}
\description{
Function to check and fix broken geometries in simple features polygon objects
}
\examples{
# Assignment of a single variable (sums)
\dontrun{
data(clea_deu2009)
out_1 <- fix_geom(clea_deu2009)
}
}
