% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_gn.R
\name{geocode_gn}
\alias{geocode_gn}
\title{Batch geocode addresses with GeoNames}
\usage{
geocode_gn(query, country_name = NULL, country_iso3 = NULL,
  str_meth = "lv", match_all = FALSE, details = FALSE, ncores = 1,
  verbose = TRUE)
}
\arguments{
\item{query}{Address(es) or place name(s) to be geocoded. Character string or vector.}

\item{country_name}{Name of country. Must supply either \code{country_name} or \code{country_iso3}. Character string.}

\item{country_iso3}{ISO 3166-1 alpha-3 country code. Must supply either \code{country_name} or \code{country_iso3}. Character string.}

\item{str_meth}{String distance metric (\code{"osa"}, \code{"lv"}, \code{"dl"}, \code{"hamming"}, \code{"lcs"}, \code{"qgram"}, \code{"cosine"}, \code{"jaccard"}, \code{"jw"}, \code{"soundex"}). Default is Levenshtein distance (\code{"lv"}). See \link[stringdist]{stringdist-metrics}. Character string.}

\item{match_all}{Attempt to match all queries? Default is \code{FALSE}. Logical.}

\item{details}{Should detailed results be returned? Default is \code{FALSE}. Logical.}

\item{ncores}{Number of cores to use for parallel socket cluster. Default is 1. Numeric.}

\item{verbose}{Print status messages and progress? Default is \code{TRUE}. Logical.}
}
\value{
A \code{data.table} object. If \code{details=FALSE}, contains fields
\itemize{
 \item{"query". }{User-supplied address query(ies). Character string.}
 \item{"geonameid". }{GeoNames ID. Character string.}
 \item{"address". }{GeoNames place name. Character string.}
 \item{"longitude". }{Horizontal coordinate. Numeric.}
 \item{"latitude". }{Vertical coordinate. Numeric.}
 }
If \code{details=TRUE}, contains additional fields
\itemize{
 \item{"country_iso3". }{ISO 3166-1 alpha-3 country code. Character string.}
 \item{"country_name". }{Country name Character string.}
 \item{"strdist". }{String distance. Numeric.}
 \item{"str_meth". }{String distance metric. Character string.}
 \item{"match_type". }{Type of match. Character string.}
 }
}
\description{
Function to find geographic coordinates of addresses and place names, using GeoNames gazetteer data.
}
\details{
This function geocodes one or multiple addresses by computing string distances between the user-supplied address query(ies) and place names (including alternate, historical and non-English spellings) in the GeoNames' free place name gazetteer data (\url{http://www.geonames.org/}). Returns geographic coordinates and other information for the matched place name. Three types of string matches, along with
\itemize{
 \item{"perfect". }{Perfect string match (e.g. "Frankfurt am Mein" -- "Frankfurt am Mein").}
 \item{"partial". }{Partial string match (e.g. "Frankfurt" -- "Frankfurt am Mein").}
 \item{"fuzzy". }{Approximate string match (e.g. "Frnkfurt" -- "Frankfurt am Mein").}
 }
Note that -- while capabale of geocoding a single address at a time -- this function is optimized for batch geocoding of multiple addresses at once. For single addresses, we recommend using \code{\link[SUNGEO]{geocode_osm}}.
}
\examples{
# Geocode an address
\dontrun{
geocode_gn("Chisinau", country_name = "Moldova")
}
# Return detailed results
\dontrun{
geocode_gn("Chisinau", country_name = "Moldova", details = TRUE)
}
# Return detailed results for multiple addresses, with progress reports
\dontrun{
geocode_gn(query = c("Chisinau","Buiucani, Chisinau","Chisinau centru"),
           country_name = "Moldova", details = TRUE, verbose = TRUE)
}
}
