% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point2poly_krig.R
\name{point2poly_krig}
\alias{point2poly_krig}
\title{Point-to-polygon interpolation, Ordinary Kriging}
\usage{
point2poly_krig(pointz, polyz, polyz2 = NULL, varz = NULL,
  cellsize = 25000, messagez = "")
}
\arguments{
\item{pointz}{Source points layer. \code{sf} object.}

\item{polyz}{Destination polygon layer. Must have identical CRS to \code{pointz}. \code{sf} object.}

\item{polyz2}{Optional polygon layer defining study region, used for identification of optimal planar projection. Must have identical CRS to \code{pointz}. \code{sf} object.}

\item{varz}{Names of numeric variable(s) to be interpolated from source points layer to destination polygons. Character string or vector of character strings.}

\item{cellsize}{Target cell size to be used in Kriging estimation. Default is 25 km. Numeric.}

\item{messagez}{Optional message to be printed during Kriging estimation. Character string.}
}
\value{
\code{sf} polygon object, with variables from \code{pointz} interpolated to the geometries of \code{polyz}. For each variable "x" in \code{varz}, two estimates are generated, with suffixes
\itemize{
 \item{"_kr". }{Mean of predicted values within each destination polygon.}
 \item{"_kr_sd". }{Average standard deviation of predicted values within each destination polygon.}
}
}
\description{
Function for interpolating values from a source points layer to an overlapping destination polygon layer, using Ordinary Kriging with automatic variogram fitting
}
\details{
This function performs Ordinary Kriging, autmatically selecting a variogram model with the smallest residual sum of squares with the sample variogram.

Unlike other available point-to-polygon interpolation techniques, this function currently only accepts numeric variables in \code{varz} and does not support interpolation of character strings.
}
\examples{
# Ordinary Kriging with one variable
\dontrun{
data(hex_05_deu)
data(clea_deu2009_pt)
out_1 <- point2poly_krig(pointz = clea_deu2009_pt,
                         polyz = hex_05_deu,
                         varz = "to1")
plot(out_1["to1_kr"])
plot(out_1["to1_kr_sd"])
}

# Ordinary Kriging with multiple variable
\dontrun{
out_2 <- point2poly_krig(pointz = clea_deu2009_pt,
                         polyz = hex_05_deu,
                         varz = c("to1","pvs1_margin"))
}
}
