% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUNGEO.R
\docType{package}
\name{SUNGEO}
\alias{SUNGEO}
\title{\code{SUNGEO}}
\description{
Sub-National Geospatial Data Archive System: Geoprocessing Toolkit
}
\details{
See the README on
[GitHub](https://github.com/zhukovyuri/SUNGEO#readme)
}
