% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpw4_deu2010}
\alias{gpw4_deu2010}
\title{Population count raster for Germany, 2010.}
\format{
class       : SpatRaster
dimensions  : 186, 220, 1  (nrow, ncol, nlyr)
resolution  : 0.04166667, 0.04166667  (x, y)
extent      : 5.875, 15.04167, 47.29167, 55.04167  (xmin, xmax, ymin, ymax)
coord. ref. : lon/lat WGS 84 (EPSG:4326)
source(s)   : memory
name        : gpw_v4_population_count_rev11_2010_2pt5_min
min value   :                                        0.00
max value   :                                    92915.66
}
\source{
Gridded Population of the World (GPW) v4: Population Count, v4.11 <doi:10.7927/H4JW8BX5>.
}
\usage{
gpw4_deu2010
}
\description{
2.5 arc-minute resolution raster of estimates of human population (number of persons per pixel),
consistent with national censuses and population registers, for the year 2010.
}
\keyword{datasets}
