% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUNGEO-package.R
\docType{package}
\name{SUNGEO-package}
\alias{SUNGEO-package}
\title{SUNGEO: Sub-National Geospatial Data Archive: Geoprocessing Toolkit}
\description{
Tools for integrating spatially-misaligned GIS datasets. Part of the Sub-National Geospatial Data Archive System.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.sungeo.org/}
  \item \url{https://github.com/zhukovyuri/SUNGEO}
}

}
\keyword{internal}
